feat: Caught page exit event and prevented prompting user to switch network to correct one when on a different tab
Added an event listener for the `blur` event to detect when the user is leaving the current tab or closing the browser. When this event is triggered, the code now checks if the user is currently on a different tab and, if so, cancels the default behavior of prompting the user to switch to the correct network. This ensures that the user is not interrupted unnecessarily and can continue using the current network without interruption. And `focus` event to detect when the user is come back.